requirements.txt
psycopg2-binary
azure-servicebus
python-dotenv